requirements.txt
certifi==14.05.14
pocket==0.3.5
python-wordpress-xmlrpc==2.3
requests==2.4.0
wsgiref==0.1.2
